translation-app/backend/requirements.txt
Django==5.2
djangorestframework==3.15.1
django-cors-headers==4.4.0
torch==2.0.0
torchtext==0.15.0
transformers==4.21.0
spacy==3.7.0
nepalitokenizers==0.0.2